[webkit] HTML5 canvas recording rendering optimization.
Currently, the canvas is being rendered on the WebCore thread into a
buffer that uses the default screen dimensions for calculating its
size. The reason is that the true screen dimensions aren't known until
we reach the UI thread in preparation for rendering to the screen. To
avoid creating canvas buffers that are larger than necessary and
require scaling before being composited, we convert the canvas buffer
into a SkPicture that gets played back on the UI thread.
We only enable the recording canvas for animations and we convert back
to a bitmap when necessary (pixel read backs, etc)

Allows for turning on/off recording canvas through a property