webkit: Fix incorrect values in tiles range calculation
When calculating the range for the tiles, left, top, right and bottom
values should be used. For baseContent, width and height can be used for
right and bottom since a PictureSet always starts from offset (0,0).
However for the viewport tile bounds, left and top could be non-zero.
Hence, the right and bottom values should be used in the calculation and
not the width and height.

CRs-Fixed: 357231

Change-Id: I77a3e7364dd25d91fdce8161aaf129ccbaa45cdc
(cherry picked from commit c4649911e6086704191b484b7bdef0a4a8f97c4c)